+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.0.0
 
+- Upgrade to react-admin v5
 - The `<LoginPage>` will no longer automatically redirect to the app if the user is already authenticated.
 
 ## 2.3.0